+++ store/src/main/java/com/secret/store/generic/MarketProviderDataStore.java
@@ -7,7 +7,7 @@ import com.secret.model.marketprovider.MarketProviderData;
 
 public abstract class MarketProviderDataStore {
   public abstract List<MarketProviderData> all() throws Exception;
-  public abstract Optional<MarketProviderData> findById(Integer id) throws Exception;
+  public abstract Optional<MarketProviderData> findById(Long id) throws Exception;
 
   public abstract void save(MarketProviderData value) throws Exception;
 }
